Create functions to create linked list and node
I created the base functions to create a linked list and to create new
nodes that can be added to the list. Additionally, I added a series of
functions to the linked list to manipulate it and get information from
it. These functions will be noted in an updated version of the README.